moved global message component
