+++ app/(public)/components/layout/global-message.tsx
