+++ app/(public)/components/layout/header.tsx
@@ -3,7 +3,7 @@ import Lockup from "@/components/patterns/lockup";
 import MainMenu from "@/components/menu/main-menu";
 import IdentityBar from "@/components/layout/identity-bar";
 import {getResource} from "@/lib/drupal/get-resource";
-import GlobalMessage from "@/components/patterns/global-message";
+import GlobalMessage from "@/components/layout/global-message";
 import {DrupalMenuLinkContent} from "next-drupal";
 import {getMenu} from "@/lib/drupal/get-menu";
 
@@ -24,7 +24,9 @@ const Header = async () => {
     <>
       <IdentityBar/>
 
-      {globalMessage?.[0]?.su_global_msg_enabled && <GlobalMessage configPage={globalMessage[0]}/>}
+      {globalMessage?.[0]?.su_global_msg_enabled &&
+          <GlobalMessage configPage={globalMessage[0]}/>
+      }
 
       <header className="su-shadow-lg su-relative su-sticky su-top-0 su-bg-white su-z-20">
         <div className="su-pt-20 su-bg-white su-max-w-1500 su-mx-auto su-px-40 2xl:su-px-0 lg:su-flex su-justify-between su-relative su-z-20 lg:su-z-10">
